Move worker resource files to a separate module
