+++ crates/tako/src/internal/common/resources/request.rs
@@ -4,7 +4,7 @@ use std::fmt;
 use crate::internal::common::error::DsError;
 use crate::internal::common::resources::{NumOfNodes, ResourceAmount, ResourceId};
 
-use crate::internal::worker::allocator::ResourceAllocator;
+use crate::internal::worker::resources::allocator::ResourceAllocator;
 use smallvec::SmallVec;
 use std::time::Duration;
