+++ crates/tako/src/internal/worker/mod.rs
@@ -1,5 +1,4 @@
 pub(crate) mod hwmonitor;
-pub(crate) mod pool;
 mod reactor;
 pub(crate) mod rpc;
 pub mod rqueue;
@@ -7,10 +6,9 @@ pub mod state;
 pub mod task;
 pub mod taskenv;
 
-pub mod allocator;
 pub mod comm;
 pub mod configuration;
-pub(crate) mod counts;
+pub(crate) mod resources;
 
 #[cfg(test)]
 mod test_util;
