+++ crates/tako/src/internal/worker/resources/mod.rs
@@ -1,0 +1,3 @@
+pub mod allocator;
+pub mod counts;
+pub mod pool;
